+++ src/main/java/com/unascribed/yttr/content/block/lazor/IRLazorBeamBlock.java
@@ -6,6 +6,7 @@ import net.minecraft.block.Block;
 import net.minecraft.block.BlockRenderType;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.ShapeContext;
+import net.minecraft.block.Waterloggable;
 import net.minecraft.block.piston.PistonBehavior;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
@@ -22,7 +23,7 @@ import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
 
-public class IRLazorBeamBlock extends AbstractLazorBlock {
+public class IRLazorBeamBlock extends AbstractLazorBlock implements Waterloggable {
 
 	public static final BooleanProperty WATERLOGGED = Properties.WATERLOGGED;