+++ src/main/java/com/unascribed/yttr/content/block/lazor/LazorBeamBlock.java
@@ -9,6 +9,7 @@ import net.minecraft.block.BlockEntityProvider;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.ShapeContext;
 import net.minecraft.block.SideShapeType;
+import net.minecraft.block.Waterloggable;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.block.entity.BlockEntityType;
@@ -34,7 +35,7 @@ import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockView;
 import net.minecraft.world.World;
 
-public class LazorBeamBlock extends AbstractColoredLazorBlock implements BlockEntityProvider {
+public class LazorBeamBlock extends AbstractColoredLazorBlock implements BlockEntityProvider, Waterloggable {
 
 	public static final BooleanProperty WATERLOGGED = Properties.WATERLOGGED;