+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 * Haemopal is now a piece of your soul
 * Continuous Platforms are now broken with *left*-click, and break instantly if not immortal
 * You can now replace your soul with soup
+* Fixed lazor emitters being half waterloggable